% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{doublePlot}
\alias{doublePlot}
\title{doublePlot}
\usage{
doublePlot(baseData, colorVec, variables, yearLim = c(-Inf, Inf),
  plotStyle = dplotStyl)
}
\description{
doublePlot
}
\author{
Roland Hollos
}
\keyword{internal}
