% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilWaterTest.R
\name{soilWaterTest}
\alias{soilWaterTest}
\title{soilWaterTest}
\usage{
soilWaterTest(year, yearDay, iniFile, epcFile = NULL, savePlot = NULL)
}
\arguments{
\item{iniFile}{The name of the inifile}

\item{epcFile}{The name of the epcfile}
}
\description{
soilWaterTest
}
