% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{AgroTest}
\alias{AgroTest}
\title{This is just a test function for AgroMo-Lite}
\usage{
AgroTest()
}
\value{
It depends on the export parameter.
}
\description{
This function runs the BBGC-MuSo model and reads in its outputfile in a very structured way, and after that plot the results automaticly
}
\author{
Roland Hollos
}
